Add serialize_packed_array
